+++ packages/core/src/generators/react/generator.ts
@@ -1,12 +1,10 @@
 import dedent from 'dedent';
 import json5 from 'json5';
-import { camelCase, upperFirst } from 'lodash';
 import { format } from 'prettier/standalone';
 import { TranspilerGenerator } from '../../types/transpiler';
 import { collectCss } from '../../helpers/styles/collect-css';
 import { createMitosisNode } from '../../helpers/create-mitosis-node';
 import { fastClone } from '../../helpers/fast-clone';
-import { filterEmptyTextNodes } from '../../helpers/filter-empty-text-nodes';
 import { getRefs } from '../../helpers/get-refs';
 import { getPropsRef } from '../../helpers/get-props-ref';
 import {
@@ -15,7 +13,6 @@ import {
 } from '../../helpers/get-state-object-string';
 import { gettersToFunctions } from '../../helpers/getters-to-functions';
 import { handleMissingState } from '../../helpers/handle-missing-state';
-import { isValidAttributeName } from '../../helpers/is-valid-attribute-name';
 import { mapRefs } from '../../helpers/map-refs';
 import { processHttpRequests } from '../../helpers/process-http-requests';
 import { processTagReferences } from '../../helpers/process-tag-references';
@@ -28,34 +25,20 @@ import {
   runPreCodePlugins,
   runPreJsonPlugins,
 } from '../../modules/plugins';
-import { selfClosingTags } from '../../parsers/jsx';
 import { MitosisComponent } from '../../types/mitosis-component';
-import { ForNode, MitosisNode } from '../../types/mitosis-node';
 import { hasContext } from '../helpers/context';
 import { collectReactNativeStyles } from '../react-native';
 import { collectStyledComponents } from '../../helpers/styles/collect-styled-components';
 import { hasCss } from '../../helpers/styles/helpers';
 import { checkHasState } from '../../helpers/state';
 import { ToReactOptions } from './types';
-import {
-  getUseStateCode,
-  processHookCode,
-  updateStateSetters,
-  updateStateSettersInCode,
-} from './state';
-import { processBinding } from './helpers';
+import { getUseStateCode, processHookCode, updateStateSetters } from './state';
+import { closeFrag, openFrag, processBinding, wrapInFragment } from './helpers';
 import hash from 'hash-sum';
-import { getForArguments } from '../../helpers/nodes/for';
 import { createSingleBinding } from '../../helpers/bindings';
+import { blockToReact } from './blocks';
 
 export const contextPropDrillingKey = '_context';
-
-const openFrag = (options: ToReactOptions) => getFragment('open', options);
-const closeFrag = (options: ToReactOptions) => getFragment('close', options);
-function getFragment(type: 'open' | 'close', options: ToReactOptions) {
-  const tagName = options.preact ? 'Fragment' : '';
-  return type === 'open' ? `<${tagName}>` : `</${tagName}>`;
-}
 
 /**
  * If the root Mitosis component only has 1 child, and it is a `Show`/`For` node, then we need to wrap it in a fragment.
@@ -64,260 +47,6 @@ function getFragment(type: 'open' | 'close', options: ToReactOptions) {
  */
 const isRootSpecialNode = (json: MitosisComponent) =>
   json.children.length === 1 && ['Show', 'For'].includes(json.children[0].name);
-
-const wrapInFragment = (json: MitosisComponent | MitosisNode) => json.children.length !== 1;
-
-const NODE_MAPPERS: {
-  [key: string]: (
-    json: MitosisNode,
-    options: ToReactOptions,
-    component: MitosisComponent,
-    parentSlots?: any[],
-  ) => string;
-} = {
-  Slot(json, options, component, parentSlots) {
-    const slotName = json.bindings.name?.code || json.properties.name;
-
-    const hasChildren = json.children.length;
-
-    const renderChildren = () => {
-      const childrenStr = json.children
-        ?.map((item) => blockToReact(item, options, component))
-        .join('\n')
-        .trim();
-      /**
-       * Ad-hoc way of figuring out if the children defaultProp is:
-       * - a JSX element, e.g. `<div>foo</div>`
-       * - a JS expression, e.g. `true`, `false`
-       * - a string, e.g. `'Default text'`
-       *
-       * and correctly wrapping it in quotes when appropriate.
-       */
-      if (childrenStr.startsWith(`<`) && childrenStr.endsWith(`>`)) {
-        return childrenStr;
-      } else if (['false', 'true', 'null', 'undefined'].includes(childrenStr)) {
-        return childrenStr;
-      } else {
-        return `"${childrenStr}"`;
-      }
-    };
-
-    if (!slotName) {
-      // TODO: update MitosisNode for simple code
-      const key = Object.keys(json.bindings).find(Boolean);
-      if (key && parentSlots) {
-        const propKey = camelCase('Slot' + key[0].toUpperCase() + key.substring(1));
-        parentSlots.push({ key: propKey, value: json.bindings[key]?.code });
-        return '';
-      }
-
-      if (hasChildren) {
-        component.defaultProps = {
-          ...(component.defaultProps || {}),
-          children: {
-            code: renderChildren(),
-            type: 'property',
-          },
-        };
-      }
-      return `{${processBinding('props.children', options)}}`;
-    }
-
-    let slotProp = processBinding(slotName as string, options).replace('name=', '');
-
-    if (!slotProp.startsWith('props.slot')) {
-      slotProp = `props.slot${upperFirst(camelCase(slotProp))}`;
-    }
-
-    if (hasChildren) {
-      component.defaultProps = {
-        ...(component.defaultProps || {}),
-        [slotProp.replace('props.', '')]: {
-          code: renderChildren(),
-          type: 'property',
-        },
-      };
-    }
-    return `{${slotProp}}`;
-  },
-  Fragment(json, options, component) {
-    const wrap = wrapInFragment(json);
-    return `${wrap ? getFragment('open', options) : ''}${json.children
-      .map((item) => blockToReact(item, options, component))
-      .join('\n')}${wrap ? getFragment('close', options) : ''}`;
-  },
-  For(_json, options, component) {
-    const json = _json as ForNode;
-    const wrap = wrapInFragment(json);
-    const forArguments = getForArguments(json).join(', ');
-    return `{${processBinding(
-      json.bindings.each?.code as string,
-      options,
-    )}?.map((${forArguments}) => (
-      ${wrap ? openFrag(options) : ''}${json.children
-      .filter(filterEmptyTextNodes)
-      .map((item) => blockToReact(item, options, component))
-      .join('\n')}${wrap ? closeFrag(options) : ''}
-    ))}`;
-  },
-  Show(json, options, component) {
-    const wrap = wrapInFragment(json);
-    return `{${processBinding(json.bindings.when?.code as string, options)} ? (
-      ${wrap ? openFrag(options) : ''}${json.children
-      .filter(filterEmptyTextNodes)
-      .map((item) => blockToReact(item, options, component))
-      .join('\n')}${wrap ? closeFrag(options) : ''}
-    ) : ${!json.meta.else ? 'null' : blockToReact(json.meta.else as any, options, component)}}`;
-  },
-};
-
-const ATTTRIBUTE_MAPPERS: { [key: string]: string } = {
-  spellcheck: 'spellCheck',
-  autocapitalize: 'autoCapitalize',
-  autocomplete: 'autoComplete',
-  for: 'htmlFor',
-};
-
-// TODO: Maybe in the future allow defining `string | function` as values
-const BINDING_MAPPERS: {
-  [key: string]:
-    | string
-    | ((key: string, value: string, options?: ToReactOptions) => [string, string]);
-} = {
-  ref(ref, value, options) {
-    const regexp = /(.+)?props\.(.+)( |\)|;|\()?$/m;
-    if (regexp.test(value)) {
-      const match = regexp.exec(value);
-      const prop = match?.[2];
-      if (prop) {
-        return [ref, prop];
-      }
-    }
-    return [ref, value];
-  },
-  innerHTML(_key, value) {
-    return ['dangerouslySetInnerHTML', `{__html: ${value.replace(/\s+/g, ' ')}}`];
-  },
-  ...ATTTRIBUTE_MAPPERS,
-};
-
-export const blockToReact = (
-  json: MitosisNode,
-  options: ToReactOptions,
-  component: MitosisComponent,
-  parentSlots?: any[],
-) => {
-  if (NODE_MAPPERS[json.name]) {
-    return NODE_MAPPERS[json.name](json, options, component, parentSlots);
-  }
-
-  if (json.properties._text) {
-    const text = json.properties._text;
-    if (options.type === 'native' && text.trim().length) {
-      return `<Text>${text}</Text>`;
-    }
-    return text;
-  }
-  if (json.bindings._text?.code) {
-    const processed = processBinding(json.bindings._text.code as string, options);
-    if (options.type === 'native') {
-      return `<Text>{${processed}}</Text>`;
-    }
-    return `{${processed}}`;
-  }
-
-  let str = '';
-
-  str += `<${json.name} `;
-
-  for (const key in json.properties) {
-    const value = (json.properties[key] || '').replace(/"/g, '&quot;').replace(/\n/g, '\\n');
-
-    if (key === 'class') {
-      str = `${str.trim()} className="${value}" `;
-    } else if (BINDING_MAPPERS[key]) {
-      const mapper = BINDING_MAPPERS[key];
-      if (typeof mapper === 'function') {
-        const [newKey, newValue] = mapper(key, value, options);
-        str += ` ${newKey}={${newValue}} `;
-      } else {
-        str += ` ${BINDING_MAPPERS[key]}="${value}" `;
-      }
-    } else {
-      if (isValidAttributeName(key)) {
-        str += ` ${key}="${(value as string).replace(/"/g, '&quot;')}" `;
-      }
-    }
-  }
-
-  for (const key in json.bindings) {
-    const value = String(json.bindings[key]?.code);
-
-    if (key === 'css' && value.trim() === '{}') {
-      continue;
-    }
-
-    const useBindingValue = processBinding(value, options);
-
-    if (json.bindings[key]?.type === 'spread') {
-      str += ` {...(${value})} `;
-    } else if (key.startsWith('on')) {
-      const { arguments: cusArgs = ['event'] } = json.bindings[key]!;
-      str += ` ${key}={(${cusArgs.join(',')}) => ${updateStateSettersInCode(
-        useBindingValue,
-        options,
-      )} } `;
-    } else if (key.startsWith('slot')) {
-      // <Component slotProjected={<AnotherComponent />} />
-      str += ` ${key}={${value}} `;
-    } else if (key === 'class') {
-      str += ` className={${useBindingValue}} `;
-    } else if (BINDING_MAPPERS[key]) {
-      const mapper = BINDING_MAPPERS[key];
-      if (typeof mapper === 'function') {
-        const [newKey, newValue] = mapper(key, useBindingValue, options);
-        str += ` ${newKey}={${newValue}} `;
-      } else {
-        str += ` ${BINDING_MAPPERS[key]}={${useBindingValue}} `;
-      }
-    } else {
-      if (isValidAttributeName(key)) {
-        str += ` ${key}={${useBindingValue}} `;
-      }
-    }
-  }
-
-  if (selfClosingTags.has(json.name)) {
-    return str + ' />';
-  }
-
-  // Self close by default if no children
-  if (!json.children.length) {
-    str += ' />';
-    return str;
-  }
-
-  // TODO: update MitosisNode for simple code
-  const needsToRenderSlots: any[] = [];
-  let childrenNodes = '';
-  if (json.children) {
-    childrenNodes = json.children
-      .map((item) => blockToReact(item, options, component, needsToRenderSlots))
-      .join('\n');
-  }
-  if (needsToRenderSlots.length) {
-    needsToRenderSlots.forEach(({ key, value }) => {
-      str += ` ${key}={${value}} `;
-    });
-  }
-  str = str.trim() + '>';
-
-  if (json.children) {
-    str += childrenNodes;
-  }
-
-  return str + `</${json.name}>`;
-};
 
 const getRefsString = (json: MitosisComponent, refs: string[], options: ToReactOptions) => {
   let hasStateArgument = false;
