+++ packages/core/src/generators/react/helpers.ts
@@ -1,3 +1,5 @@
+import { MitosisComponent } from '../../types/mitosis-component';
+import { MitosisNode } from '../../types/mitosis-node';
 import { stripStateAndPropsRefs } from '../../helpers/strip-state-and-props-refs';
 
 import { ToReactOptions } from './types';
@@ -12,3 +14,11 @@ export const processBinding = (str: string, options: ToReactOptions) => {
     includeProps: false,
   });
 };
+
+export const openFrag = (options: ToReactOptions) => getFragment('open', options);
+export const closeFrag = (options: ToReactOptions) => getFragment('close', options);
+export function getFragment(type: 'open' | 'close', options: ToReactOptions) {
+  const tagName = options.preact ? 'Fragment' : '';
+  return type === 'open' ? `<${tagName}>` : `</${tagName}>`;
+}
+export const wrapInFragment = (json: MitosisComponent | MitosisNode) => json.children.length !== 1;
